@@ -15,8 +15,8 @@
   "license": "UNLICENSED",
   "dependencies": {
     "@ahmadnassri/spawn-promise": "^1.2.0",
-    "@balena/jellyfish-client-sdk": "^3.4.1",
-    "@balena/jellyfish-jellyscript": "^3.0.190",
+    "@balena/jellyfish-client-sdk": "^5.5.0",
+    "@balena/jellyfish-jellyscript": "^4.9.11",
     "@types/rewire": "^2.5.28",
     "dockerode": "^3.3.0",
     "fast-json-patch": "^3.0.0-1",
@@ -24,11 +24,11 @@
     "lodash": "^4.17.21",
     "memory-streams": "^0.1.3",
     "node-rsa": "^1.1.1",
-    "tar": "^6.1.0"
+    "tar": "^6.1.6"
   },
   "devDependencies": {
     "@balena/lint": "^5.4.2",
-    "@balena/jellyfish-types": "^0.5.116",
+    "@balena/jellyfish-types": "^0.8.13",
     "@types/dockerode": "^3.2.3",
     "@types/lockfile": "^1.0.1",
     "@types/node": "^14.14.45",
@@ -40,6 +40,6 @@
     "rewire": "^5.0.0",
     "ts-jest": "^26.5.6",
     "ts-node": "^9.1.1",
-    "typescript": "^4.2.4"
+    "typescript": "^4.3.5"
   }
 }
